✨ Feat: 사이드바 기능 개선 및 구조 변경
