+++ src/layouts/Header.tsx
@@ -1,6 +1,6 @@
 import { useEffect, useRef, useState } from "react";
 import Icon from "../assets/icons/Icon";
-import Sidebar from "./Sidebar";
+import Sidebar from "./Sidebar/Sidebar";
 import Notification from "../components/notification/notification";
 
 const Header: React.FC = () => {
@@ -36,7 +36,7 @@ const Header: React.FC = () => {
   return (
     <>
       <header className="w-full h-[84px] top-0 left-0 fixed px-10 py-[21px] z-20 bg-blue-7  ">
-        <div className="max-w-[1440px] mx-auto flex justify-between items-center">
+        <div className="max-w-[1280px] mx-auto flex justify-between items-center">
           <a href="/" aria-label="홈페이지로 이동">
             <h1 className="font-dm italic text-blue-1 text-2xl/[38px]">
               On culture
